+++ src/tiled/propertiesdock.cpp
@@ -95,9 +95,9 @@ PropertiesDock::PropertiesDock(QWidget *parent)
 
     mPropertyBrowser->setContextMenuPolicy(Qt::CustomContextMenu);
     connect(mPropertyBrowser, &PropertyBrowser::customContextMenuRequested,
-                this, &PropertiesDock::showContextMenu);
-    connect(mPropertyBrowser, SIGNAL(currentItemChanged(QtBrowserItem*)),
-            SLOT(currentItemChanged(QtBrowserItem*)));
+            this, &PropertiesDock::showContextMenu);
+    connect(mPropertyBrowser, &PropertyBrowser::currentItemChanged,
+            this, &PropertiesDock::updateActions);
 
     retranslateUi();
 }
@@ -115,6 +115,11 @@ void PropertiesDock::setDocument(Document *document)
                 SLOT(currentObjectChanged(Object*)));
         connect(document, SIGNAL(editCurrentObject()),
                 SLOT(bringToFront()));
+
+        connect(document, &Document::propertyAdded,
+                this, &PropertiesDock::updateActions);
+        connect(document, &Document::propertyRemoved,
+                this, &PropertiesDock::updateActions);
 
         currentObjectChanged(document->currentObject());
     } else {
@@ -144,18 +149,12 @@ static bool isPartOfTileset(const Object *object)
     }
 }
 
-static bool isAddedByType(Object *object, const QString &name)
+static bool anyObjectHasProperty(const QList<Object*> &objects, const QString &name)
 {
-    if (!object || object->typeId() != Object::MapObjectType)
-        return false;
-
-    const QString objectType = static_cast<MapObject*>(object)->type();
-    const ObjectTypes objectTypes = Preferences::instance()->objectTypes();
-    for (const ObjectType &type : objectTypes) {
-        if (type.name == objectType)
-            return type.defaultProperties.contains(name);
+    for (Object *obj : objects) {
+        if (obj->hasProperty(name))
+            return true;
     }
-
     return false;
 }
 
@@ -171,16 +170,15 @@ void PropertiesDock::currentObjectChanged(Object *object)
     mActionAddProperty->setEnabled(enabled);
 }
 
-void PropertiesDock::currentItemChanged(QtBrowserItem *item)
+void PropertiesDock::updateActions()
 {
+    QtBrowserItem *item = mPropertyBrowser->currentItem();
     bool isCustomProperty = mPropertyBrowser->isCustomPropertyItem(item);
     bool editingTileset = mDocument && mDocument->type() == Document::TilesetDocumentType;
     bool isTileset = isPartOfTileset(mPropertyBrowser->object());
-    bool enabled = !isTileset || editingTileset;
-
-    bool addedByType = item && isAddedByType(mPropertyBrowser->object(),
-                                     item->property()->propertyName());
-    bool canModify = isCustomProperty && enabled && !addedByType;
+    bool canModify = isCustomProperty &&
+            (!isTileset || editingTileset) &&
+            anyObjectHasProperty(mDocument->currentObjects(), item->property()->propertyName());
 
     mActionRemoveProperty->setEnabled(canModify);
     mActionRenameProperty->setEnabled(canModify);
@@ -223,15 +221,6 @@ void PropertiesDock::removeProperty()
 
     const QString name = item->property()->propertyName();
     QUndoStack *undoStack = mDocument->undoStack();
-    QList<QtBrowserItem *> items = item->parent()->children();
-    if (items.count() > 1) {
-        int currentItemIndex = items.indexOf(item);
-        if (item == items.last()) {
-            mPropertyBrowser->setCurrentItem(items.at(currentItemIndex - 1));
-        } else {
-            mPropertyBrowser->setCurrentItem(items.at(currentItemIndex + 1));
-        }
-    }
     undoStack->push(new RemoveProperty(mDocument,
                                        mDocument->currentObjects(),
                                        name));